Fixed bugs, and updated to pytorch 1.4 version

+++ scripts/gen_mnist.py
@@ -41,8 +41,10 @@ data_dir = path.join('data', task)
 input_dir = path.join(data_dir, 'MNIST', 'processed')
 output_dir = path.join(data_dir, 'pt', metric_dir)
 output_input_dir = path.join(output_dir, 'input')
-utils.rmdir(output_input_dir); utils.mkdir(output_input_dir)
 output_gt_dir = path.join(output_dir, 'gt')
+if arg.v == 0:
+    utils.rmdir(output_input_dir); utils.mkdir(output_input_dir)
+    utils.rmdir(output_gt_dir); utils.mkdir(output_gt_dir)
 
 
 # mnist data
@@ -87,38 +89,38 @@ def process_batch(states, batch_id):
             if states[o][0] < appear_interval: # wait for interval frames 
                 states[o][0] = states[o][0] + 1
             elif states[o][0] == appear_interval: # allow birth
-                if unif[t][o] < birth_prob: # birth
+                if unif[t][o].item() < birth_prob: # birth
                     # shape and appearance
-                    data_ind = data_id[t][o]
-                    scale = scales[t][o]
-                    ratio = ratios[t][o]
+                    data_ind = data_id[t][o].item()
+                    scale = scales[t][o].item()
+                    ratio = ratios[t][o].item()
                     h_, w_ = round(h * scale * ratio), round(w * scale / ratio)
                     data_patch = data[data_ind]
                     # data_patch = utils.imresize(data[data_ind], h_, w_).unsqueeze(2)
                     # pose
-                    direction = direction_id[t][o]
+                    direction = direction_id[t][o].item()
                     position = position_id[t][o]
                     x1, y1, x2, y2 = None, None, None, None
                     if direction == 0:
-                        x1 = position[0]
+                        x1 = position[0].item()
                         y1 = m
-                        x2 = position[1]
+                        x2 = position[1].item()
                         y2 = H - 1 - m
                     elif direction == 1:
-                        x1 = position[0]
+                        x1 = position[0].item()
                         y1 = H - 1 - m
-                        x2 = position[1]
+                        x2 = position[1].item()
                         y2 = m
                     elif direction == 2:
                         x1 = m
-                        y1 = position[0]
+                        y1 = position[0].item()
                         x2 = W - 1 - m
-                        y2 = position[1]
+                        y2 = position[1].item()
                     else:
                         x1 = W - 1 - m
-                        y1 = position[0]
+                        y1 = position[0].item()
                         x2 = m
-                        y2 = position[1]
+                        y2 = position[1].item()
                     theta = math.atan2(y2 - y1, x2 - x1)
                     vx = velocity * math.cos(theta)
                     vy = velocity * math.sin(theta)
